requirements.txt
gspread==5.10.0
python-dotenv==1.0.0